requirements.txt
joblib
numpy
torch
umap-learn
pandas
seaborn
colorcet
matplotlib
scikit-learn
sentence-transformers
pillow
nltk